[chore] remove calls to validate the configuration, now done in the config builder
