+++ exporter/mezmoexporter/factory_test.go
@@ -55,9 +55,7 @@ func TestIngestUrlMustConform(t *testing.T) {
 	cfg.IngestURL = "https://example.com:8088/services/collector"
 	cfg.IngestKey = "1234-1234"
 
-	params := exportertest.NewNopCreateSettings()
-	_, err := createLogsExporter(context.Background(), params, cfg)
-	assert.Error(t, err, `"ingest_url" must end with "/otel/ingest/rest"`)
+	assert.Error(t, cfg.Validate(), `"ingest_url" must end with "/otel/ingest/rest"`)
 }
 
 func TestCreateLogsExporter(t *testing.T) {
@@ -74,32 +72,4 @@ func TestCreateLogsExporterNoConfig(t *testing.T) {
 	params := exportertest.NewNopCreateSettings()
 	_, err := createLogsExporter(context.Background(), params, nil)
 	assert.Error(t, err)
-}
-
-func TestCreateLogsExporterInvalidEndpoint(t *testing.T) {
-	cfg := createDefaultConfig().(*Config)
-	cfg.IngestURL = "urn:something:12345"
-	params := exportertest.NewNopCreateSettings()
-	_, err := createLogsExporter(context.Background(), params, cfg)
-	assert.Error(t, err)
-}
-
-func TestCreateInstanceViaFactory(t *testing.T) {
-	factory := NewFactory()
-
-	cfg := factory.CreateDefaultConfig().(*Config)
-	cfg.IngestURL = "https://example.com:8088/otel/ingest/rest"
-	cfg.IngestKey = "1234-1234"
-	params := exportertest.NewNopCreateSettings()
-	exp, err := factory.CreateLogsExporter(context.Background(), params, cfg)
-	assert.NoError(t, err)
-	assert.NotNil(t, exp)
-	assert.NoError(t, exp.Shutdown(context.Background()))
-
-	// Set values that don't have a valid default.
-	cfg.IngestURL = "https://example.com"
-	cfg.IngestKey = "testToken"
-	exp, err = factory.CreateLogsExporter(context.Background(), params, cfg)
-	assert.Error(t, err)
-	assert.Nil(t, exp)
 }
