+++ receiver/statsdreceiver/factory_test.go
@@ -17,16 +17,12 @@ package statsdreceiver
 import (
 	"context"
 	"testing"
-	"time"
 
-	"github.com/lightstep/go-expohisto/structure"
 	"github.com/stretchr/testify/assert"
 	"go.opentelemetry.io/collector/component"
 	"go.opentelemetry.io/collector/component/componenttest"
 	"go.opentelemetry.io/collector/consumer/consumertest"
 	"go.opentelemetry.io/collector/receiver/receivertest"
-
-	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/statsdreceiver/protocol"
 )
 
 type testHost struct {
@@ -56,101 +52,6 @@ func TestCreateReceiver(t *testing.T) {
 	tReceiver, err := createMetricsReceiver(context.Background(), params, cfg, consumertest.NewNop())
 	assert.NoError(t, err)
 	assert.NotNil(t, tReceiver, "receiver creation failed")
-}
-
-func TestCreateReceiverWithConfigErr(t *testing.T) {
-	cfg := &Config{
-		AggregationInterval: -1,
-		TimerHistogramMapping: []protocol.TimerHistogramMapping{
-			{StatsdType: "timing", ObserverType: "gauge"},
-		},
-	}
-	receiver, err := createMetricsReceiver(
-		context.Background(),
-		receivertest.NewNopCreateSettings(),
-		cfg,
-		consumertest.NewNop(),
-	)
-	assert.Error(t, err, "aggregation_interval must be a positive duration")
-	assert.Nil(t, receiver)
-
-}
-
-func TestCreateReceiverWithHistogramConfigError(t *testing.T) {
-	for _, maxSize := range []int32{structure.MaximumMaxSize + 1, -1, -structure.MaximumMaxSize} {
-		cfg := &Config{
-			AggregationInterval: 20 * time.Second,
-			TimerHistogramMapping: []protocol.TimerHistogramMapping{
-				{
-					StatsdType:   "timing",
-					ObserverType: "histogram",
-					Histogram: protocol.HistogramConfig{
-						MaxSize: maxSize,
-					},
-				},
-			},
-		}
-		receiver, err := createMetricsReceiver(
-			context.Background(),
-			receivertest.NewNopCreateSettings(),
-			cfg,
-			consumertest.NewNop(),
-		)
-		assert.Error(t, err)
-		assert.Contains(t, err.Error(), "histogram max_size out of range")
-		assert.Nil(t, receiver)
-	}
-}
-
-func TestCreateReceiverWithHistogramGoodConfig(t *testing.T) {
-	for _, maxSize := range []int32{structure.MaximumMaxSize, 0, 2} {
-		cfg := &Config{
-			AggregationInterval: 20 * time.Second,
-			TimerHistogramMapping: []protocol.TimerHistogramMapping{
-				{
-					StatsdType:   "timing",
-					ObserverType: "histogram",
-					Histogram: protocol.HistogramConfig{
-						MaxSize: maxSize,
-					},
-				},
-			},
-		}
-		receiver, err := createMetricsReceiver(
-			context.Background(),
-			receivertest.NewNopCreateSettings(),
-			cfg,
-			consumertest.NewNop(),
-		)
-		assert.NoError(t, err)
-		assert.NotNil(t, receiver)
-		assert.NoError(t, receiver.Start(context.Background(), &testHost{t: t}))
-		assert.NoError(t, receiver.Shutdown(context.Background()))
-	}
-}
-
-func TestCreateReceiverWithInvalidHistogramConfig(t *testing.T) {
-	cfg := &Config{
-		AggregationInterval: 20 * time.Second,
-		TimerHistogramMapping: []protocol.TimerHistogramMapping{
-			{
-				StatsdType:   "timing",
-				ObserverType: "gauge",
-				Histogram: protocol.HistogramConfig{
-					MaxSize: 100,
-				},
-			},
-		},
-	}
-	receiver, err := createMetricsReceiver(
-		context.Background(),
-		receivertest.NewNopCreateSettings(),
-		cfg,
-		consumertest.NewNop(),
-	)
-	assert.Error(t, err)
-	assert.Contains(t, err.Error(), "histogram configuration requires observer_type: histogram")
-	assert.Nil(t, receiver)
 }
 
 func TestCreateMetricsReceiverWithNilConsumer(t *testing.T) {
